Move types to transport
