+++ dds/src/rtps/message_receiver.rs
@@ -8,8 +8,9 @@ use super::{
     stateful_reader::RtpsStatefulReader,
     stateful_writer::RtpsStatefulWriter,
     stateless_reader::RtpsStatelessReader,
-    types::{GuidPrefix, Locator, ProtocolVersion, VendorId, GUIDPREFIX_UNKNOWN},
+    types::{ProtocolVersion, VendorId},
 };
+use crate::transport::types::{GuidPrefix, Locator, GUIDPREFIX_UNKNOWN};
 
 pub struct MessageReceiver {
     source_version: ProtocolVersion,
