+++ dds/src/rtps/message_sender.rs
@@ -5,11 +5,10 @@ use network_interface::{Addr, NetworkInterface, NetworkInterfaceConfig};
 
 use super::{
     messages::overall_structure::{RtpsMessageHeader, RtpsMessageWrite, Submessage},
-    types::{
-        GuidPrefix, Locator, ProtocolVersion, VendorId, LOCATOR_KIND_UDP_V4, LOCATOR_KIND_UDP_V6,
-        PROTOCOLVERSION_2_4, VENDOR_ID_S2E,
-    },
+    types::{ProtocolVersion, VendorId, PROTOCOLVERSION_2_4, VENDOR_ID_S2E},
 };
+
+use crate::transport::types::{GuidPrefix, Locator, LOCATOR_KIND_UDP_V4, LOCATOR_KIND_UDP_V6};
 
 pub struct MessageSender {
     protocol_version: ProtocolVersion,
