+++ dds/src/rtps/messages/submessages/data_frag.rs
@@ -1,4 +1,7 @@
-use crate::rtps::messages::submessage_elements::SerializedDataFragment;
+use crate::{
+    rtps::messages::submessage_elements::SerializedDataFragment,
+    transport::types::{EntityId, SequenceNumber},
+};
 
 use super::super::super::{
     error::{RtpsError, RtpsErrorKind, RtpsResult},
@@ -10,7 +13,6 @@ use super::super::super::{
         submessage_elements::ParameterList,
         types::{FragmentNumber, SubmessageFlag, SubmessageKind},
     },
-    types::{EntityId, SequenceNumber},
 };
 use std::io::Write;
 
@@ -218,11 +220,11 @@ impl Submessage for DataFragSubmessage {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::rtps::{
-        messages::{
+    use crate::{
+        rtps::messages::{
             overall_structure::write_submessage_into_bytes_vec, submessage_elements::Parameter,
         },
-        types::{USER_DEFINED_READER_GROUP, USER_DEFINED_READER_NO_KEY},
+        transport::types::{USER_DEFINED_READER_GROUP, USER_DEFINED_READER_NO_KEY},
     };
 
     #[test]
