+++ dds/src/rtps/messages/submessages/gap.rs
@@ -1,3 +1,5 @@
+use crate::transport::types::{EntityId, SequenceNumber};
+
 use super::super::super::{
     error::RtpsResult,
     messages::{
@@ -8,7 +10,6 @@ use super::super::super::{
         submessage_elements::SequenceNumberSet,
         types::SubmessageKind,
     },
-    types::{EntityId, SequenceNumber},
 };
 use std::io::Write;
 
@@ -83,9 +84,9 @@ impl Submessage for GapSubmessage {
 
 #[cfg(test)]
 mod tests {
-    use crate::rtps::{
-        messages::overall_structure::write_submessage_into_bytes_vec,
-        types::{USER_DEFINED_READER_GROUP, USER_DEFINED_READER_NO_KEY},
+    use crate::{
+        rtps::messages::overall_structure::write_submessage_into_bytes_vec,
+        transport::types::{USER_DEFINED_READER_GROUP, USER_DEFINED_READER_NO_KEY},
     };
 
     #[test]
