+++ dds/src/rtps/messages/submessages/heartbeat_frag.rs
@@ -1,3 +1,5 @@
+use crate::transport::types::{EntityId, SequenceNumber};
+
 use super::super::super::{
     error::RtpsResult,
     messages::{
@@ -7,7 +9,6 @@ use super::super::super::{
         },
         types::{Count, FragmentNumber, SubmessageKind},
     },
-    types::{EntityId, SequenceNumber},
 };
 use std::io::Write;
 
@@ -92,9 +93,9 @@ impl Submessage for HeartbeatFragSubmessage {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::rtps::{
-        messages::overall_structure::write_submessage_into_bytes_vec,
-        types::{USER_DEFINED_READER_GROUP, USER_DEFINED_READER_NO_KEY},
+    use crate::{
+        rtps::messages::overall_structure::write_submessage_into_bytes_vec,
+        transport::types::{USER_DEFINED_READER_GROUP, USER_DEFINED_READER_NO_KEY},
     };
 
     #[test]
