+++ dds/src/rtps/writer_proxy.rs
@@ -1,4 +1,4 @@
-use crate::transport::types::ReliabilityKind;
+use crate::transport::types::{EntityId, Guid, Locator, ReliabilityKind, SequenceNumber};
 
 use super::{
     message_sender::MessageSender,
@@ -11,7 +11,6 @@ use super::{
         },
         types::Count,
     },
-    types::{EntityId, Guid, Locator, SequenceNumber},
 };
 
 use std::{cmp::max, collections::HashMap, sync::Arc};
